requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==0.3.0
google-api-python-client==2.143.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
langgraph==0.2.40
langchain-core==0.3.15
pydantic==2.9.2
python-dotenv==1.0.1
apscheduler==3.10.4
jinja2==3.1.4
sqlalchemy==2.0.36
requests==2.31.0
